internal_crash_catcher: test_app: allowing all processes to attach
Problem is that test_app and gdb are both children of CrashDialog,
and it would be required to pass gdb's pid to test_app.
That would require changes in CrashDialog which could make
DEVELOPER_BUILD path too inconsistent with normal way of
CrashDialog usage (where crashing app is parent of CrashDialog
which is parent of gdb).