cleanups + signals postponing
Without postpone there is a problem when rows insertion
causes row deletion (setupComboBox discovers that after
insertion, there is no place for combo, so row with combo is removed).

I'm not quite sure what's the problem, but it is possible
that base View doesn't have time to handle these two events in right
order.