removing Type from GroupInfo
Type is property of group not property of photo
and GroupInfo is for photo properties.

There was also an ambiguity when updating photo -
a group's member: Type of group could be changed for it
and there was no effect (sql backend was intentionally ignoring it)
as, again, it was property of group.

Type was moved to Group.
For now LazyTreeItemDelegate is broken as it needs group type.